Update an location

Updates an existing location.

updateLocation(locationId, name, streetAddress, extendedAddress, locality,
  region, postalCode, phone, website, hoursOfOperationExplicit,
  hoursOfOperationNotes, tourInfo, timezoneId, latitude, longitude,
  isPrimary, inPlanning, isClosed, openToPublic, locationType,
  countryIsoCode)

Arguments

locationId
The locationId

name
Nickname for the location

streetAddress
Street address

extendedAddress
Extended address, such as suite or apartment number

locality
Locality, or city, where the location is at

region
Region, also known as state or province

postalCode
Postal code

phone
Phone number for the location

website
Location-specific website, if different from the brewery

hoursOfOperationExplicit
Explict breakdown of the hours of operation. Should be in format "ddd-hh:mm(am/pm)-hh:mm(am/pm)". Multiple days can be passed with the given format, separated by commas. Example: 'mon-8:00am-10:00pm,tue-9:00am-1:00pm' is valid.

hoursOfOperationNotes
Additional, non-time-related notes about the hours of operation.

tourInfo
Brewery tour information for visitors of the location

timezoneId
Timezone ID for the location. This value is automatically updated based on the latitude and longitude of the location, but can be passed as well. Example: America/New_York

latitude
Latitude for the location

longitude
Longitude for the location

isPrimary
Whether or not this is the primary location for the brewery Y N Default

inPlanning
Whether or not this location is in planning Y N Default

isClosed
Whether or not this location is currently closed Y N Default

openToPublic
Whether or not this location is open to the public Y N Default

locationType
Type of location, such as micro brewery or brew pub micro macro nano prewpub production office tasting restaurant cidery meadery

countryIsoCode
Required Country that the location is located in

Value

none